Modules work!
You have to make sure the person exists in the people.yml otherwise
things go a little crazy right now, but that's not really a big deal I
don't think.

This is really just content for now, we'll get more later, and a nice
template.

The only ones that work currently are catch and servlet.